pyproject.toml
[project]
name = "rishta-app-with-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
